dashboard/backend/requirements.txt
Flask
pandas
plotly
psycopg2-binary 
sqlalchemy 
pulp 
country_converter
python-gnupg
requests
flask-cors
python-dotenv